+++ src/components/__tests__/HeadingBackground.cy.js
@@ -1,4 +1,4 @@
-import HeadingBackground from '../HeadingBackground.vue';
+import HeadingBackground from '../homepage/HeadingBackground.vue';
 
 const title = 'Join the Bike to Work community in your city';
