+++ src/components/__tests__/ListCardFollow.cy.js
@@ -1,5 +1,5 @@
 import { colors } from 'quasar';
-import ListCardFollow from '../ListCardFollow.vue';
+import ListCardFollow from '../homepage/ListCardFollow.vue';
 import { i18n } from '../../boot/i18n';
 
 // mocks
