+++ src/components/global/DialogCard.vue
