+++ src/components/homepage/CardChallenge.vue
@@ -28,7 +28,7 @@
 import { defineComponent } from 'vue';
 
 // types
-import { CardChallenge as CardChallengeType, ConfigGlobal } from './types';
+import { CardChallenge as CardChallengeType, ConfigGlobal } from '../types';
 
 // config
 const rideToWorkByBikeConfig: ConfigGlobal = JSON.parse(
