+++ src/components/homepage/NewsletterItem.vue
@@ -24,7 +24,7 @@
 import { defineComponent, computed } from 'vue';
 
 // types
-import { NewsletterItem as NewsletterItemType } from './types';
+import { NewsletterItem as NewsletterItemType } from '../types';
 
 export default defineComponent({
   name: 'NewsletterItem',
